requirements.txt
pydantic>=2.7
pydantic-ai>=0.0.14
anthropic>=0.40.0
anyio>=4.0
tavily-python>=0.5.0
asyncpg>=0.29
openai>=1.0
pytest-asyncio>=0.23
